requirements.txt
fastapi
uvicorn
python-multipart
httpx
pydantic-settings
openai
pymupdf
pypdfium2
pdfplumber
pillow
